refactor: split root dropping into its own package
